Refactor debugger code to its own file
